# docs/swagger.yaml
basePath: /v1
definitions:
...
  handlers.LoginRequestBody:
    properties:
      email:
        example: example@gmail.com
        type: string
      password:
        example: P@ssw0rd
        type: string
    type: object
...
  handlers.Response:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  handlers.TokensResponseBody:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
  models.Comment:
    properties:
      body:
        example: fuga aut est delectus...
        type: string
      email:
        example: lindsey@caitlyn.net
        type: string
      id:
        example: 302
        type: integer
      name:
        example: quia voluptatem...
        type: string
      postId:
        example: 61
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: Artem Kostenko
    url: https://github.com/aerosystems
  description: Simple REST API for CRUD operations with Comments & Posts enities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: NIX Trainee 5-6-7-8 tasks
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Comment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: get all comments
      tags:
      - comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: delete comment by ID
      tags:
      - comments
    get:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: get comment by ID
      tags:
      - comments
    patch:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: particular update comment by ID
      tags:
      - comments
    post:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: create comment by ID
      tags:
      - comments
...
  /users/login:
    post:
      consumes:
      - application/json
      - text/xml
      description: |-
        Password should contain:
        - minimum of one small case letter
        - minimum of one upper case letter
        - minimum of one digit
        - minimum of one special character
        - minimum 8 characters length
      parameters:
      - description: raw request body
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequestBody'
      produces:
      - application/json
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.TokensResponseBody'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: login user by credentionals
      tags:
      - users
  /users/logout:
    post:
      consumes:
      - application/json
      - text/xml
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/xml
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: logout user
      tags:
      - users
...
swagger: "2.0"